tests: make model tests load existing pkd config (wip)
