--- tests/pipeline/nodes/model/csrnetv1/test_csrnet.yml
@@ -1,14 +1,0 @@
-input: ["img"]
-output: ["density_map", "count"]
-
-model_type: sparse # sparse or dense
-weights_parent_dir: null
-weights: {
-    model_subdir: csrnet,
-    blob_file: csrnet.zip,
-    model_file: {
-        dense: shanghaitech-part-a,
-        sparse: shanghaitech-part-b
-    }
-}
-width: 640
