--- tests/pipeline/nodes/model/efficientdet_d04/test_efficientdet.yml
@@ -1,23 +1,0 @@
-input: ['img']
-output: ['bboxes', 'bbox_labels', 'bbox_scores']
-
-model_type: 0 # 0-4
-weights_parent_dir: null
-weights: {
-    model_subdir: efficientdet,
-    blob_file: efficientdet.zip,
-    classes_file: coco_90.json,
-    model_file: {
-        0: efficientdet-d0.pb,
-        1: efficientdet-d1.pb,
-        2: efficientdet-d2.pb,
-        3: efficientdet-d3.pb,
-        4: efficientdet-d4.pb,
-    }
-}
-size: [512, 640, 768, 896, 1024, 1280, 1408]
-num_classes: 90
-score_threshold: 0.3
-detect_ids: [0]
-MODEL_NODES:
-    { inputs: [x:0], outputs: [Identity:0, Identity_1:0, Identity_2:0] }
