--- tests/pipeline/nodes/model/hrnetv1/test_hrnet.yml
@@ -1,12 +1,0 @@
-input: ['img', 'bboxes']
-output: ['keypoints', 'keypoint_scores', 'keypoint_conns']
-
-weights_parent_dir: null
-weights: {
-    model_subdir: hrnet,
-    blob_file: hrnet.zip,
-    model_file: hrnet_frozen.pb
-}
-resolution: { height: 192, width: 256 }
-score_threshold: 0.1
-MODEL_NODES: { inputs: [x:0], outputs: [Identity:0] }
