--- tests/pipeline/nodes/model/jdev1/test_jde.yml
@@ -1,21 +1,0 @@
-input: ["img"]
-output: ["bboxes", "bbox_labels", "bbox_scores", "obj_attrs"]
-
-optional_inputs: ["mot_metadata"]
-
-weights_parent_dir: null
-weights:
-  {
-    model_subdir: jde,
-    blob_file: jde.zip,
-    model_file: { 864x480: jde_864x480_uncertainty-stripped-half.pt },
-    config_file: { 864x480: yolov3_864x480.cfg },
-  }
-
-model_type: 864x480
-iou_threshold: 0.5
-nms_threshold: 0.4
-score_threshold: 0.5
-
-min_box_area: 200
-track_buffer: 30
